process_18_12_(3)

+++ src/main/java/org/example/dao_repositories_implements/LessonVisitingDAOImpl.java
@@ -16,16 +16,27 @@ public class LessonVisitingDAOImpl implements LessonVisitingDAO {
     }
 
     @Override
-    public String AddLessonVisiting(int lessonId, ArrayList<String> students) {
-
-        LessonVisiting lessonVisiting = new LessonVisiting(0, lessonId, students);
-        dataBase.addLessonVisiting(lessonVisiting);
-        return "Посещаемость урока добавлена!";
+    public String AddLessonVisiting(int lessonId, ArrayList<String> students) throws Exception {
+        LessonVisiting result = dataBase.getLessonVisiting(lessonId);
+        if (result == null){
+            LessonVisiting lessonVisiting = new LessonVisiting(0, lessonId, students);
+            dataBase.addLessonVisiting(lessonVisiting);
+            return "Посещаемость урока добавлена!";
+        }
+        else{
+            throw new Exception();
+        }
     }
 
     @Override
-    public String DeleteLessonVisiting(int lessonId) {
-        dataBase.deleteLessonVisiting(lessonId);
-        return "Посещаемость данного урока удалена!";
+    public String DeleteLessonVisiting(int lessonId) throws Exception {
+        LessonVisiting result = dataBase.getLessonVisiting(lessonId);
+        if (result == null){
+            dataBase.deleteLessonVisiting(lessonId);
+            return "Посещаемость данного урока удалена!";
+        }
+        else{
+            throw new Exception();
+        }
     }
 }
